requirements.txt
alembic==0.9.6
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-MongoAlchemy==0.7.2
Flask-PyMongo==0.5.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
MongoAlchemy==0.19
pymongo==2.8.1
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.0
visitor==0.1.3
Werkzeug==0.13
WTForms==2.1
